requirements.txt
alembic==1.4.3
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.10
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudinary==1.22.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.6
django-uuidfield==0.5.0
djangorestframework==3.12.2
dnspython==2.0.0
dominate==2.5.2
duplicity==0.8.12.0
email-validator==1.1.1
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==20.0.4
httpie==1.0.3
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyDrive==1.3.1
Pygments==2.3.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-editor==1.0.4
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.12
sqlparse==0.3.1
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
visitor==0.1.3
wadllib==1.3.3
Werkzeug==0.16.1
WTForms==2.3.3
xkit==0.0.0
zipp==1.0.0
